refactored page type
