+++ src/index.tsx
@@ -3,40 +3,15 @@ import ReactDOM from 'react-dom/client';
 import {RouterProvider, createBrowserRouter} from 'react-router-dom';
 import App from './App.tsx';
 import './index.css';
-import Home from './pages/Home.tsx';
-import About from './pages/About.tsx';
-import Projects from './pages/Projects.tsx';
-import Contact from './pages/Contact.tsx';
-import Articles from './pages/Articles.tsx';
-import PageNotFound from './pages/PageNotFound.tsx';
+import PageNotFound from './pages/page_not_found/PageNotFound.tsx';
+import Page from './common/types/page.type.ts'
 
 const router = createBrowserRouter([
   {
     path: '/',
     element: <App />,
     errorElement: <PageNotFound />,
-    children: [
-      {
-        path: '/',
-        element: <Home />,
-      },
-      {
-        path: '/about',
-        element: <About />,
-      },
-      {
-        path: '/projects',
-        element: <Projects />,
-      },
-      {
-        path: '/articles',
-        element: <Articles />,
-      },
-      {
-        path: '/contact',
-        element: <Contact />,
-      },
-    ],
+    children: Object.values(Page)
   },
 ]);
